pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = [
  "setuptools>=42",
  "setuptools-scm>=7",
]

[project]
name = "curviriver"
description = "Generate curvilinear mesh from river outline polygons"
readme = { file = "README.rst", content-type = "text/x-rst" }
keywords = [
  'shapely',
  'mesh',
  'curvilinear grid',
]
license = { text = "MIT" }
maintainers = [{ name = "Taher Chegini", email = "cheginit@gmail.com" }]
requires-python = ">=3.8"
classifiers = [
  'Development Status :: 3 - Alpha',
  'Intended Audience :: Science/Research',
  'License :: OSI Approved :: MIT License',
  'Operating System :: OS Independent',
  'Programming Language :: Python',
  "Programming Language :: Python :: 3 :: Only",
  'Programming Language :: Python :: 3.8',
  'Programming Language :: Python :: 3.9',
  'Programming Language :: Python :: 3.10',
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  'Topic :: Scientific/Engineering',
]
dynamic = [
  "version",
]
dependencies = [
  "cytoolz",
  "geopandas>=0.12",
  "networkx>=2.5",
  "numpy>=1.21",
  "pandas>=1",
  "pyproj>=2.2",
  "scikit-learn>=1.3",
  "scipy",
  "shapely>=2",
]
[project.optional-dependencies]
doc = [
  "ipykernel",
  "ipython",
  "ipywidgets",
  "jupyter_client",
  "jupyter_sphinx",
  "nbsphinx",
  "numpydoc",
  "pydata-sphinx-theme",
  "sphinx",
  "sphinx-autoapi",
  "sphinx-copybutton",
  "sphinx-design",
  "sphinx-favicon>=1.0.1",
  "sphinx-gallery",
  "sphinx-sitemap",
  "sphinx-togglebutton",
  "sphinxcontrib-bibtex",
  "sphinxext-opengraph",
]
parallel = [
  "joblib",
]
test = [
  "coverage[toml]",
  "pytest-cov",
  "pytest-xdist[psutil]",
]
[project.urls]
Code = "https://github.com/cheginit/curviriver"
Home = "https://curviriver.readthedocs.io"
Issues = "https://github.com/cheginit/curviriver/issues"

[tool.setuptools]
include-package-data = true
license-files = ["LICENSE"]
platforms = ["any"]
zip-safe = false

[tool.setuptools.package-data]
curviriver = ["py.typed"]

[tool.setuptools.packages]
find = {namespaces = false}

[tool.setuptools_scm]
fallback_version = "999"

[tool.black]
line-length = 100

[tool.ruff]
target-version = 'py38'
select = [
    # flake8-bugbear
    "B",
    # flake8-comprehensions
    "C4",
    # pydocstyle
    "D",
    # Error
    "E",
    # pyflakes
    "F",
    # isort
    "I",
    # flake8-implicit-str-concat
    "ISC",
    # pep8-naming
    "N",
    # pygrep-hooks
    "PGH",
    # flake8-use-pathlib
    "PTH",
    # flake8-quotes
    "Q",
    # bandit
    "S",
    # flake8-simplify
    "SIM",
    # tryceratops
    "TRY",
    # pyupgrade
    "UP",
    # Warning
    "W",
    # flake8-2020
    "YTT",
    # flake8-debugger
    "T10",
    # flake8-gettext
    "INT",
    # pylint
    "PLC",
    "PLE",
    "PLR",
    "PLW",
    # misc lints
    "PIE",
    # flake8-pyi
    "PYI",
    # tidy imports
    "TID",
    # type-checking imports
    "TCH",
    # Ruff-specific rules
    "RUF",
]

exclude = [
    "__pycache__",
    ".nox",
]

ignore = [
    "D103",
    "D105",
    "E501",
    "PLR2004",
    "PLR0913",
]
line-length = 100

[tool.ruff.flake8-bugbear]
extend-immutable-calls = [
    "chr",
    "typer.Argument",
    "typer.Option",
]

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.ruff.per-file-ignores]
"tests/*.py" = [
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    # use of "assert"
    "S101",
    # use of "exec"
    "S102",
    # possible hardcoded password
    "S106",
    # use of "eval"
    "PGH001",
    # Mutable class attributes
    "RUF012",
]

[tool.ruff.isort]
known-first-party = [
  "curviriver",
]

[tool.pytest.ini_options]
addopts = "--ignore=noxfile.py -n=auto -v --cov=curviriver --cov-report xml --durations=5"
doctest_optionflags = 'NORMALIZE_WHITESPACE IGNORE_EXCEPTION_DETAIL NUMBER'
filterwarnings = [
  "ignore:.*distutils.*",
  "ignore:.*--rsyncdir command line argument.*",
  "ignore:.*numpy.ndarray size changed.*",
  "ignore:.*invalid value encountered in cast.*",
  "ignore:.*'cgi' is deprecated.*",
]
testpaths = [
  "tests",
  "curviriver"
]

[tool.coverage.report]
exclude_lines = [
  'raise ServiceUnavailableError',
  "if TYPE_CHECKING:"
]
ignore_errors = true

[tool.coverage.run]
branch = true
omit = [
  "**/__init__.py",
]
parallel = true
source = ['curviriver']

[tool.fixit]
disable = [
    "fixit.rules:NoStringTypeAnnotation",  # single lint rule by name
]

[tool.codespell]
skip = "__pycache__,_build,.mypy_cache,.git,./htmlcov,.nox,build,*.ipynb"

[tool.pycln]
all = true
expand_stars = true
no_gitignore = false
verbose = true

[tool.pyright]
exclude = [
  ".nox/",
  "**/__pycache__",
  "**/__init__.py",
  "tests/"
]
include = ["curviriver"]
reportMissingTypeStubs = false
reportUnknownArgumentType = false
reportUnknownLambdaType = false
reportUnknownMemberType = false
reportUnknownParameterType = false
reportUnknownVariableType = false
reportUnnecessaryIsInstance = false
typeCheckingMode = "strict"

[tool.refurb]
python_version = "3.8"
quiet = true
